backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==3.12.0
cloudinary==1.28.0
gunicorn==20.1.0
python-jose==3.3.0
requests==2.26.0
dnspython==2.1.0 